pytest-health/instrumentation/requirements.txt
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation==0.46b0
pytest-xdist==3.6.1
pytest==8.3.2
requests==2.32.3
opentelemetry-container-distro==0.2.0